Game 1 (Moving Without Restrictions)
This Is the First 'Game'/'Challenge', Named "Moving Without Restrictions". The purpose of this game is giving the player freedom to move without any restrictions in any direction on the 3D plane.
In addition, the camera should follow the player. Should The camera be locked or should it have some calculation logic, or should it simply be controlled by input is up to the developer, who wants to try this challenge.
Best Of Luck.
PS. how you write code doesn't matter it can be with physics or without them it is up to you.